Refacto + refuse file when process is running
