+++ gridcapa-task-manager-app/src/test/java/com/farao_community/farao/gridcapa/task_manager/app/service/EventHandlerTest.java
@@ -1,20 +1,17 @@
 /*
- * Copyright (c) 2021, RTE (http://www.rte-france.com)
+ * Copyright (c) 2023, RTE (http://www.rte-france.com)
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.farao_community.farao.gridcapa.task_manager.app;
+package com.farao_community.farao.gridcapa.task_manager.app.service;
 
-import com.farao_community.farao.gridcapa.task_manager.api.TaskStatusUpdate;
+import com.farao_community.farao.gridcapa.task_manager.app.TaskRepository;
+import com.farao_community.farao.gridcapa.task_manager.app.TaskUpdateNotifier;
 import com.farao_community.farao.gridcapa.task_manager.app.entities.ProcessEvent;
 import com.farao_community.farao.gridcapa.task_manager.app.entities.Task;
-import com.farao_community.farao.minio_adapter.starter.MinioAdapter;
-import com.farao_community.farao.minio_adapter.starter.MinioAdapterConstants;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
@@ -23,16 +20,13 @@ import org.springframework.cloud.stream.function.StreamBridge;
 import java.time.OffsetDateTime;
 import java.util.UUID;
 
-import static com.farao_community.farao.gridcapa.task_manager.api.TaskStatus.RUNNING;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 /**
- * @author Joris Mancini {@literal <joris.mancini at rte-france.com>}
+ * @author Theo Pascoli {@literal <theo.pascoli at rte-france.com>}
  */
 @SpringBootTest
-@ExtendWith(MockitoExtension.class)
-class TaskManagerTest {
-    private static final String INPUT_FILE_GROUP_VALUE = MinioAdapterConstants.DEFAULT_GRIDCAPA_INPUT_GROUP_METADATA_VALUE;
+class EventHandlerTest {
 
     @Autowired
     private TaskUpdateNotifier taskUpdateNotifier;
@@ -40,22 +34,15 @@ class TaskManagerTest {
     @MockBean
     private StreamBridge streamBridge; // Useful to avoid AMQP connection that would fail
 
-    @MockBean
-    private MinioAdapter minioAdapter;
+    @Autowired
+    private EventHandler eventHandler;
 
     @Autowired
     private TaskRepository taskRepository;
 
-    @Autowired
-    private ProcessFileRepository processFileRepository;
-
-    @Autowired
-    private TaskManager taskManager;
-
     @AfterEach
     void cleanDatabase() {
         taskRepository.deleteAll();
-        processFileRepository.deleteAll();
     }
 
     @Test
@@ -71,7 +58,7 @@ class TaskManagerTest {
                 "  \"message\": \"Hello from backend\",\n" +
                 "  \"serviceName\": \"GRIDCAPA\" \n" +
                 "}";
-        taskManager.handleTaskEventUpdate(logEvent);
+        eventHandler.handleTaskEventUpdate(logEvent);
         Task updatedTask = taskRepository.findByTimestamp(taskTimestamp).orElseThrow();
         assertEquals(1, updatedTask.getProcessEvents().size());
         ProcessEvent event = updatedTask.getProcessEvents().iterator().next();
@@ -94,36 +81,12 @@ class TaskManagerTest {
                 "  \"serviceName\": \"GRIDCAPA\" ,\n" +
                 "  \"eventPrefix\": \"STEP-1\" \n" +
                 "}";
-        taskManager.handleTaskEventUpdate(logEvent);
+        eventHandler.handleTaskEventUpdate(logEvent);
         Task updatedTask = taskRepository.findByTimestamp(taskTimestamp).orElseThrow();
         assertEquals(1, updatedTask.getProcessEvents().size());
         ProcessEvent event = updatedTask.getProcessEvents().iterator().next();
         assertEquals(OffsetDateTime.parse("2021-12-30T16:31:33.030Z"), event.getTimestamp());
         assertEquals("INFO", event.getLevel());
         assertEquals("[STEP-1] : Hello from backend", event.getMessage());
-    }
-
-    @Test
-    void handleTaskStatusUpdateFromMessagesTest() {
-        OffsetDateTime taskTimestamp = OffsetDateTime.parse("2021-10-01T21:00Z");
-        Task task = new Task(taskTimestamp);
-        taskRepository.save(task);
-
-        taskManager.handleTaskStatusUpdate(new TaskStatusUpdate(task.getId(), RUNNING));
-
-        Task updatedTask = taskRepository.findByTimestamp(taskTimestamp).orElseThrow();
-        assertEquals(RUNNING, updatedTask.getStatus());
-    }
-
-    @Test
-    void handleTaskStatusUpdateFromApiTest() {
-        OffsetDateTime taskTimestamp = OffsetDateTime.parse("2021-10-01T21:00Z");
-        Task task = new Task(taskTimestamp);
-        taskRepository.save(task);
-
-        taskManager.handleTaskStatusUpdate(taskTimestamp, RUNNING);
-
-        Task updatedTask = taskRepository.findByTimestamp(taskTimestamp).orElseThrow();
-        assertEquals(RUNNING, updatedTask.getStatus());
     }
 }
